python/requirements.txt
# This file is generated by pipreqs.
# Note, the flightsql-dbapi cannot be resolved by pipreqs and it's added manually.
flightsql-dbapi==0.2.1
pandas==2.2.3
protobuf==5.28.2
pyarrow==17.0.0
